+++ alicloud/import_alicloud_cs_kubernetes_test.go
@@ -12,7 +12,7 @@ func TestAccAlicloudCSKubernetes_import(t *testing.T) {
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckContainerClusterDestroy,
+		CheckDestroy: testAccCheckKubernetesClusterDestroy,
 		Steps: []resource.TestStep{
 			resource.TestStep{
 				Config: testAccContainerKubernetes_basic,