+++ alicloud/resource_alicloud_cs_kubernetes_test.go
@@ -3,8 +3,11 @@ package alicloud
 import (
 	"testing"
 
+	"fmt"
+
 	"github.com/denverdino/aliyungo/cs"
 	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform/terraform"
 )
 
 func TestAccAlicloudCSKubernetes_basic(t *testing.T) {
@@ -18,14 +21,14 @@ func TestAccAlicloudCSKubernetes_basic(t *testing.T) {
 		IDRefreshName: "alicloud_cs_kubernetes.k8s",
 
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckContainerClusterDestroy,
+		CheckDestroy: testAccCheckKubernetesClusterDestroy,
 		Steps: []resource.TestStep{
 			resource.TestStep{
 				Config: testAccContainerKubernetes_basic,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckContainerClusterExists("alicloud_cs_kubernetes.k8s", &k8s),
-					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "worker_number", "3"),
-					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "name", "terraform-test-for-k8s-basic"),
+					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "worker_number", "1"),
+					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "name", "testAccContainerKubernetes-basic"),
 					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "master_nodes.#", "3"),
 					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "master_disk_category", "cloud_ssd"),
 					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "worker_disk_size", "50"),
@@ -49,13 +52,13 @@ func TestAccAlicloudCSKubernetes_autoVpc(t *testing.T) {
 		IDRefreshName: "alicloud_cs_kubernetes.k8s",
 
 		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckContainerClusterDestroy,
+		CheckDestroy: testAccCheckKubernetesClusterDestroy,
 		Steps: []resource.TestStep{
 			resource.TestStep{
 				Config: testAccContainerKubernetes_autoVpc,
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckContainerClusterExists("alicloud_cs_kubernetes.k8s", &k8s),
-					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "worker_number", "3"),
+					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "worker_number", "1"),
 					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "master_nodes.#", "3"),
 					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "worker_disk_category", "cloud_ssd"),
 					resource.TestCheckResourceAttr("alicloud_cs_kubernetes.k8s", "master_disk_size", "50"),
@@ -68,12 +71,35 @@ func TestAccAlicloudCSKubernetes_autoVpc(t *testing.T) {
 	})
 }
 
-const testAccContainerKubernetes_basic = `
+func testAccCheckKubernetesClusterDestroy(s *terraform.State) error {
+	client := testAccProvider.Meta().(*AliyunClient).csconn
 
-provider "alicloud" {
-	region="cn-shanghai"
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "alicloud_cs_kubernetes" {
+			continue
+		}
+
+		cluster, err := client.DescribeCluster(rs.Primary.ID)
+
+		if err != nil {
+			if NotFoundError(err) || IsExceptedError(err, ErrorClusterNotFound) {
+				continue
+			}
+			return err
+		}
+
+		if cluster.ClusterID != "" {
+			return fmt.Errorf("Error container cluster %s still exists.", rs.Primary.ID)
+		}
+	}
+
+	return nil
 }
 
+const testAccContainerKubernetes_basic = `
+variable "name" {
+	default = "testAccContainerKubernetes-basic"
+}
 data "alicloud_zones" main {
   available_resource_creation = "VSwitch"
 }
@@ -85,7 +111,7 @@ data "alicloud_instance_types" "default" {
 }
 
 resource "alicloud_vpc" "foo" {
-  name = "tf_test_image"
+  name = "${var.name}"
   cidr_block = "10.1.0.0/21"
 }
 
@@ -96,12 +122,12 @@ resource "alicloud_vswitch" "foo" {
 }
 
 resource "alicloud_cs_kubernetes" "k8s" {
-  name = "terraform-test-for-k8s-basic"
+  name = "${var.name}"
   vswitch_id = "${alicloud_vswitch.foo.id}"
   new_nat_gateway = true
   master_instance_type = "${data.alicloud_instance_types.default.instance_types.0.id}"
   worker_instance_type = "${data.alicloud_instance_types.default.instance_types.0.id}"
-  worker_number = 3
+  worker_number = 1
   master_disk_category  = "cloud_ssd"
   worker_disk_size = 50
   password = "Test12345"
@@ -116,7 +142,9 @@ const testAccContainerKubernetes_autoVpc = `
 provider "alicloud" {
 	region="us-west-1"
 }
-
+variable "name" {
+	default = "testAccContainerKubernetes-autoVpc"
+}
 data "alicloud_zones" main {
   available_resource_creation = "VSwitch"
 }
@@ -128,12 +156,12 @@ data "alicloud_instance_types" "default" {
 }
 
 resource "alicloud_cs_kubernetes" "k8s" {
-  name_prefix = "terraform-test-for-k8s"
+  name_prefix = "${var.name}"
   availability_zone = "${data.alicloud_zones.main.zones.0.id}"
   new_nat_gateway = true
   master_instance_type = "${data.alicloud_instance_types.default.instance_types.0.id}"
   worker_instance_type = "${data.alicloud_instance_types.default.instance_types.0.id}"
-  worker_number = 3
+  worker_number = 1
   password = "Test12345"
   pod_cidr = "172.20.0.0/16"
   service_cidr = "172.21.0.0/20"